feat(generate): update v1.0.3: + handle sync for file <= 6MB + change progress loading + add upload file
